+++ lib/twirp_rails/error_handling/error_handling.rb
@@ -1,0 +1,2 @@
+require 'twirp_rails/error_handling/base'
+require 'twirp_rails/error_handling/error_handling_factory'