+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.4 - 2020-05-05
+
+### Added
+
+- ability to translate twirp errors to exceptions and vise versa.
+
+## 0.4.3 - 2020-04-14
+
+### Fixed
+
+- autorequire all ruby files from `lib/twirp_clients` folder.
+
+## 0.4.2 - 2020-04-01
+
+### Added
+
+- `to_twirp` extension use model method unless attribute found 
+
 ## 0.4.1 - 2020-03-30
 
 ### Added